fix: pass lint
